+++ tests/test_style.py
@@ -27,3 +27,17 @@ def test_dont_write_redundant_tables():
     actual = tomli_w.dumps({"tab1": {"tab2": {"tab3": {}}}})
     expected = "[tab1.tab2.tab3]\n"
     assert actual == expected
+
+
+def test_multiline():
+    multiline_string = (
+        "This is longer than threshold!\n"
+        "Should be formatted as a multiline basic string"
+    )
+    actual = tomli_w.dumps({"ml_string": multiline_string})
+    expected = '''\
+ml_string = """
+This is longer than threshold!
+Should be formatted as a multiline basic string"""
+'''
+    assert actual == expected