+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## **unreleased**
+
+- Changed
+  - Format strings containing line breaks and that are longer than threshold value as multiline strings
+
 ## 0.1.0
 
 - Added